requirements.txt
python-telegram-bot==20.0
httpx==0.23.0
beautifulsoup4==4.11.1
python-dotenv==0.21.0
